// 前后端源代码/front-employee/src/views/sys_department/dept/utils/types.ts
interface FormItemProps {
  gender: any;
  higherDeptOptions: Record<string, unknown>[];
  parentId: number;
  name: string;
  principal: string;
  phone: string | number;
  email: string;
  sort: number;
  status: number;
  remark: string;
}
interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };